backend/requirements.txt
django-resized==1.0.2
djangorestframework==3.14.0
Pillow==9.3.0
django-cors-headers==3.13.0
